+++ forms/src/directives/numberInput/maxValueNumberValidator.directive.ts
@@ -2,7 +2,7 @@ import {NG_VALIDATORS, AbstractControl, Validator, ValidatorFn, ValidationErrors
 import {ExistingProvider, Input, forwardRef, Directive, Attribute, OnChanges, SimpleChanges} from '@angular/core';
 import {isBlank, isPresent} from '@jscrpt/common';
 
-import {Validators} from '../../misc';
+import {Validators} from '../../misc/validators';
 
 /**
  * Validator that is injected with directive MaxValueNumberValidatorDirective